fix(templates/drizzle-neon): Move Drizzle database setup to index file
