+++ templates/javascript/drizzle-neondb-sample/src/function/index.js
@@ -1,6 +1,9 @@
 import { MainRouter } from "../lib/routers/main-router.js";
 import { PageRouter } from "../lib/routers/page-router.js";
 import { ApiRouter } from "../lib/routers/api-router.js";
+import { neon } from '@neondatabase/serverless';
+import { drizzle } from 'drizzle-orm/neon-http';
+import { pgTable, text, serial } from "drizzle-orm/pg-core";
 
 async function handleRequest(request, args) {
   // to add or change a route check the route files
@@ -18,7 +21,20 @@ async function handleRequest(request, args) {
     return new Response(`You are trying something incorrect.`, { status: 200 });
   });
 
-  return mainRouter.handle(request, { args });
+  //Setup Neon connection
+  const client = neon(
+    args.connection_url ||
+    Azion.env.get("DRIZZLE_NEON_CONNECTION_URL")
+  );
+  const db = drizzle(client);
+
+  //Setup Neon table
+  const posts = pgTable('posts', {
+    id: serial('id'),
+    message: text('message'),
+  });
+
+  return mainRouter.handle(request, db, posts);
 }
 
 export { handleRequest };
